+++ LotgdFormat.Tests/src/Cleanup.cs
@@ -4,61 +4,29 @@ using LotgdFormat;
 using Xunit;
 
 public class Cleanup {
-	[Fact]
-	public void Keeps_Data() {
-		var formatter = new Formatter(new List<LotgdFormatCode> {
-			new LotgdFormatCode(){ Token = 'c', Tag = "center"}
-		});
-		formatter.AddText("`cThis is centered");
-		Assert.Equal("<center>This is centered", formatter.GetOutput());
-		Assert.Equal("<center>This is centered", formatter.GetOutput());
-	}
 
 	[Fact]
 	public void Clears_Content() {
 		var formatter = new Formatter(new List<LotgdFormatCode> {
 			new LotgdFormatCode(){ Token = 'c', Tag = "center"}
 		});
-		formatter.AddText("`cThis is centered");
-		Assert.Equal("<center>This is centered", formatter.GetOutput());
-		formatter.ClearText();
-		Assert.Equal("", formatter.GetOutput());
-		formatter.CloseOpenTags();
-		Assert.Equal("</center>", formatter.GetOutput());
+		Assert.Equal("<center>This is centered", formatter.AddText("`cThis is centered"));
+		Assert.Equal("", formatter.AddText(""));
+		Assert.Equal("</center>", formatter.CloseOpenTags());
 	}
-
-	[Fact]
-	public void Clears_All() {
-		var formatter = new Formatter(new List<LotgdFormatCode> {
-			new LotgdFormatCode(){ Token = 'c', Tag = "center"}
-		});
-		formatter.AddText("`cThis is centered");
-		Assert.Equal("<center>This is centered", formatter.GetOutput());
-		formatter.Clear();
-		Assert.Equal("", formatter.GetOutput());
-		formatter.CloseOpenTags();
-		Assert.Equal("", formatter.GetOutput());
-	}
-
 
 	[Fact]
 	public void DoubleCloseOpenTags() {
 		var formatter = new Formatter(new List<LotgdFormatCode> {
 			new LotgdFormatCode(){ Token = 'c', Tag = "center"}
 		});
-		formatter.AddText("`cThis is centered");
-		var foo = formatter.GetOutput();
+		var foo = formatter.AddText("`cThis is centered");
 		Assert.Equal("<center>This is centered", foo);
 		Assert.False(formatter.IsClear());
-		formatter.CloseOpenTags();
-		Assert.Equal(foo + "</center>", formatter.GetOutput());
-		formatter.CloseOpenTags();
-		Assert.Equal(foo + "</center>", formatter.GetOutput());
+		Assert.Equal("</center>", formatter.CloseOpenTags());
+		Assert.Equal("", formatter.CloseOpenTags());
 		formatter.Clear();
-		Assert.Equal("", formatter.GetOutput());
-		formatter.Clear();
-		Assert.False(formatter.HasContent());
-		formatter.Clear();
+		Assert.Equal("", formatter.CloseOpenTags());
 		Assert.True(formatter.IsClear());
 	}
 
@@ -69,11 +37,9 @@ public class Cleanup {
 			new LotgdFormatCode(){ Token = '$', Color = "FF0000"},
 			new LotgdFormatCode(){ Token = '@', Color = "00FF00"},
 		});
-		formatter.AddText("`$");
-		var foo = formatter.GetOutput();
-		formatter.ClearText();
-		formatter.AddText("FooBar`0");
-		foo += formatter.GetOutput();
+		;
+		var foo = formatter.AddText("`$");
+		foo += formatter.AddText("FooBar`0");
 		Assert.Equal("<span class=\"c36\">FooBar</span>", foo);
 	}
 }