+++ LotgdFormat.Tests/src/Color.cs
@@ -9,8 +9,8 @@ public class Color {
 		var formatter = new Formatter(new List<LotgdFormatCode> {
 			new LotgdFormatCode(){ Token = '@', Color = "00FF00"}
 		});
-		formatter.AddText("This is White `@this is green.");
-		string result = formatter.GetOutput();
+
+		string result = formatter.AddText("This is White `@this is green.");
 
 		Assert.Equal("This is White <span class=\"c64\">this is green.", result);
 	}
@@ -20,9 +20,8 @@ public class Color {
 		var formatter = new Formatter(new List<LotgdFormatCode> {
 			new LotgdFormatCode(){ Token = '@', Color = "00FF00"}
 		});
-		formatter.AddText("`0This is plaintext.");
-		string result = formatter.GetOutput();
 
+		string result = formatter.AddText("`0This is plaintext.");
 		Assert.Equal("This is plaintext.", result);
 	}
 
@@ -33,8 +32,7 @@ public class Color {
 			new LotgdFormatCode(){ Token = '$', Color = "FF0000"}
 		});
 		formatter.Color = false;
-		formatter.AddText("`$red `@green `$red `@green");
-		string result = formatter.GetOutput();
+		string result = formatter.AddText("`$red `@green `$red `@green");
 
 		Assert.Equal("red green red green", result);
 	}
@@ -44,13 +42,13 @@ public class Color {
 		var formatter = new Formatter(new List<LotgdFormatCode> {
 			new LotgdFormatCode(){ Token = '@', Color = "00FF00"}
 		});
-		formatter.AddText("This is White `@this is green.");
-		string result = formatter.GetOutput();
+
+		string result = formatter.AddText("This is White `@this is green.");
 
 		Assert.Equal("This is White <span class=\"c64\">this is green.", result);
 
-		formatter.AddText(" This is still green.");
-		result = formatter.GetOutput();
+
+		result += formatter.AddText(" This is still green.");
 		Assert.Equal("This is White <span class=\"c64\">this is green. This is still green.", result);
 	}
 
@@ -60,13 +58,11 @@ public class Color {
 		var formatter = new Formatter(new List<LotgdFormatCode> {
 			new LotgdFormatCode(){ Token = '@', Color = "00FF00"}
 		});
-		formatter.AddText("This is White `@this is green.");
-		string result = formatter.GetOutput();
+		string result = formatter.AddText("This is White `@this is green.");
 
 		Assert.Equal("This is White <span class=\"c64\">this is green.", result);
 
-		formatter.AddText("`0 This is no longer green.");
-		result = formatter.GetOutput();
+		result += formatter.AddText("`0 This is no longer green.");
 		Assert.Equal("This is White <span class=\"c64\">this is green.</span> This is no longer green.", result);
 	}
 
@@ -75,12 +71,10 @@ public class Color {
 		var formatter = new Formatter(new List<LotgdFormatCode> {
 			new LotgdFormatCode(){ Token = '@', Color = "00FF00"}
 		});
-		formatter.AddText("`@green");
-		string result = formatter.GetOutput();
 
+		string result = formatter.AddText("`@green");
 		Assert.Equal("<span class=\"c64\">green", result);
-		formatter.CloseOpenTags();
-		result = formatter.GetOutput();
+		result += formatter.CloseOpenTags();
 		Assert.Equal("<span class=\"c64\">green</span>", result);
 	}
 
@@ -90,8 +84,8 @@ public class Color {
 			new LotgdFormatCode(){ Token = '@', Color = "00FF00"},
 			new LotgdFormatCode(){ Token = '$', Color = "FF0000"}
 		});
-		formatter.AddText("`@green`$red");
-		string result = formatter.GetOutput();
+
+		string result = formatter.AddText("`@green`$red");
 
 		Assert.Equal("<span class=\"c64\">green</span><span class=\"c36\">red", result);
 	}
@@ -102,8 +96,8 @@ public class Color {
 			new LotgdFormatCode(){ Token = '@', Color = "00FF00"},
 			new LotgdFormatCode(){ Token = '$', Color = "FF0000"}
 		});
-		formatter.AddText("`@`0Normal text");
-		string result = formatter.GetOutput();
+
+		string result = formatter.AddText("`@`0Normal text");
 
 		Assert.Equal("Normal text", result);
 	}
@@ -115,8 +109,8 @@ public class Color {
 			new LotgdFormatCode(){ Token = '$', Color = "FF0000"},
 			new LotgdFormatCode(){ Token = 'b', Tag = "strong"}
 		});
-		formatter.AddText("`b`@`0strong text");
-		string result = formatter.GetOutput();
+
+		string result = formatter.AddText("`b`@`0strong text");
 
 		Assert.Equal("<strong>strong text", result);
 	}
@@ -130,17 +124,9 @@ public class Color {
 			new LotgdFormatCode(){ Token = 'b', Tag = "strong"}
 		});
 		formatter.AddText("`0Dummy");
-		formatter.GetOutput();
-		formatter.ClearText();
 		formatter.AddText("Dummy");
-		formatter.GetOutput();
-		formatter.ClearText();
 		formatter.AddText("`0`&Dummy`0`&");
-		formatter.GetOutput();
-		formatter.ClearText();
 		formatter.AddText("Dummy`0");
-		formatter.GetOutput();
-		formatter.ClearText();
 		Assert.True(true);
 	}
 }