+++ LotgdFormat.Tests/src/ConfigHandling.cs
@@ -9,8 +9,7 @@ public class InvalidConfig {
 		var formatter = new Formatter(new List<LotgdFormatCode> {
 			new LotgdFormatCode(){ Token = 'E', Tag = null, Color = null}
 		});
-		formatter.AddText("`EThis is defective`0");
-		string result = formatter.GetOutput();
+		string result = formatter.AddText("`EThis is defective`0");
 
 		Assert.Equal("This is defective", result);
 	}
@@ -20,8 +19,7 @@ public class InvalidConfig {
 		var formatter = new Formatter(new List<LotgdFormatCode> {
 			new LotgdFormatCode(){ Token = 'E', Tag = "b", Color = "FF0000"}
 		});
-		formatter.AddText("`EThis is not bold.`0");
-		string result = formatter.GetOutput();
+		string result = formatter.AddText("`EThis is not bold.`0");
 
 		Assert.Equal("<span class=\"c69\">This is not bold.</span>", result);
 	}
@@ -31,14 +29,11 @@ public class InvalidConfig {
 		var formatter = new Formatter(new List<LotgdFormatCode> {
 			new LotgdFormatCode(){ Token = 'H', Tag = "span", Style = "class=\"navhi\"", Privileged = true}
 		});
-		formatter.AddText("This is `Hadmin reserved.`H");
-		string result = formatter.GetOutput();
 
+		string result = formatter.AddText("This is `Hadmin reserved.`H");
 		Assert.Equal("This is admin reserved.", result);
-
 		formatter.Clear();
-		formatter.AddText("This is `Hadmin reserved.`H", isPrivileged: true);
-		string result2 = formatter.GetOutput();
+		string result2 = formatter.AddText("This is `Hadmin reserved.`H", isPrivileged: true);
 		Assert.Equal("This is <span class=\"navhi\">admin reserved.</span>", result2);
 	}
 
@@ -48,9 +43,8 @@ public class InvalidConfig {
 			new LotgdFormatCode(){ Token = '@', Color = "00FF00"},
 			new LotgdFormatCode(){ Token = '$', Color = "FF0000"}
 		});
-		formatter.AddText("regular `@green`0 `_regular");
-		string result = formatter.GetOutput();
 
+		string result = formatter.AddText("regular `@green`0 `_regular");
 		Assert.Equal("regular <span class=\"c64\">green</span> _regular", result);
 	}
 }