+++ LotgdFormat.Tests/src/Nesting.cs
@@ -10,8 +10,8 @@ public class Nesting {
 			new LotgdFormatCode(){ Token = '@', Color="00FF00"},
 			new LotgdFormatCode(){ Token = 'b', Tag="b"}
 		});
-		formatter.AddText("`@green `bbold green`0 bold`b");
-		string result = formatter.GetOutput();
+
+		string result = formatter.AddText("`@green `bbold green`0 bold`b");;
 
 		Assert.Equal("<span class=\"c64\">green <b>bold green</b></span><b> bold</b>", result);
 	}
@@ -24,13 +24,13 @@ public class Nesting {
 			new LotgdFormatCode(){ Token = 'b', Tag="b"},
 			new LotgdFormatCode(){ Token = 'n', Tag="br", SelfClosing = true}
 		});
-		formatter.AddText("`b`&with text`0`b `bbold`b");
-		string result = formatter.GetOutput();
+
+		string result = formatter.AddText("`b`&with text`0`b `bbold`b");
 
 		Assert.Equal("<b><span class=\"c38\">with text</span></b> <b>bold</b>", result);
 		formatter.Clear();
-		formatter.AddText("`b`&`0`b `bbold`b");
-		result = formatter.GetOutput();
+
+		result = formatter.AddText("`b`&`0`b `bbold`b");
 
 		Assert.Equal(" <b>bold</b>", result);
 	}